requirements.txt
opencv-python
numpy
torch
torchvision
ultralytics
transformers
pillow
